Fix buggy for() loop in cli_free_entry_all()
cli_free_entry_all() was freeing entries in the list using a pointer (p) freed and then read again to get the next entry in the list. This is logically incorrect.

The function signature has been changed to allow direct modification of the list head so that head is left with a NULL value when freeing is done.