protocol: add new DNDS result 'moreData'
This flag will be used to inform the receiver side
that a large message has been truncated. When netvirt-switch
query netvirt-ctrler to fetch the list of all the context, during
its bootstrap period, the DNDSMessage is getting too big.
Now, netvirt-switch will stay in the bootstraping state as long
as the status of every DNDSResult is 'moreData' and will exit
this state when it receive a message containing 'success'.